Fix an important bug :  how to close the individual child panel independently.
Please notice the function's argument ===> QuitUserInterface(panel);   It will return current panel ID of the child window which you want to close,  to function "hPanel = RunUserInterface()" and "DiscardPanel(hPanel)".

That is an advanced technique, make good use of the feature of function "RunUserInterface()",  Please see its help to understand the technique.